move ProxySourceService creation to Configuration class
