+++ src/main/resources/application.properties
@@ -10,12 +10,10 @@ spring.security.oauth2.authorizationserver.endpoint.jwk-set-uri=http://${keycloa
 
 logging.level.org.springframework.security=DEBUG
 
-worker.base.uri= 172.31.15.68:8080
+worker.base.uri=172.31.15.68:8080
 
 # Proxy
 proxy.filepath=ProxyConfigs.json
 proxy.webshare.api-key=1avpih9tre77omhxxvw07276m8pbzzhfdfqwat84
 proxy.webshare.api-url=https://proxy.webshare.io/api/v2/proxy/list/?mode=direct&page=1&page_size=25&valid=true
-
-proxy.source.type=http
-
+proxy.source.type=url
