--- src/test/java/com/gfl/client/ClientApplicationTests.java
@@ -1,13 +1,0 @@
-package com.gfl.client;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ClientApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
