+++ src/test/java/com/gfl/client/model/ScenarioRequestTest.java
@@ -42,11 +42,11 @@ class ScenarioRequestTest {
         List<StepRequest> steps = Collections.singletonList(new StepRequest("click", "button"));
         List<StepRequest> step = Collections.singletonList(new StepRequest(null, null));
         return Stream.of(
-                new Object[]{new ScenarioRequest(), 5},
+                new Object[]{new ScenarioRequest(), 3},
                 new Object[]{new ScenarioRequest("Scenario", "https://example.com", steps), 0},
                 new Object[]{new ScenarioRequest("", "https://example.com", steps), 2},
                 new Object[]{new ScenarioRequest("Scenario", "", steps), 2},
-                new Object[]{new ScenarioRequest("Scenario", "https://example.com", step), 4},
+                new Object[]{new ScenarioRequest("Scenario", "https://example.com", step), 2},
                 new Object[]{new ScenarioRequest("Scenario", "", new ArrayList<>()), 3}
         );
     }
