+++ src/test/resources/application-test.properties
@@ -1,0 +1,1 @@
+proxy.source.type=
